feat: make ProgressModal more reusable
